requirements.txt
sounddevice==0.4.6
numpy==1.24.3
google-cloud-speech==2.21.0
python-dotenv==1.0.0